LPS-58194 SF - move control panel application service to product navigation
