--- modules/apps/product-menu/product-menu-control-panel/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Product Menu Control Panel Service
-Bundle-SymbolicName: com.liferay.product.menu.control.panel.service
-Bundle-Version: 1.0.0
-Include-Resource: classes
