+++ modules/apps/product-navigation/product-navigation-control-panel/.classpath
