+++ modules/apps/product-navigation/product-navigation-control-panel/.project
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <projectDescription>
-	<name>product-menu-control-panel-master</name>
+	<name>product-navigation-control-panel-master</name>
 	<comment></comment>
 	<projects></projects>
 	<buildSpec>
