+++ modules/apps/product-navigation/product-navigation-control-panel/snapshot.properties
