+++ modules/apps/product-navigation/product-navigation-control-panel/src/com/liferay/product/navigation/control/panel/application/list/ConfigurationPanelCategory.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.product.menu.control.panel.application.list;
+package com.liferay.product.navigation.control.panel.application.list;
 
-import com.liferay.application.list.BaseControlPanelEntryPanelApp;
-import com.liferay.application.list.PanelApp;
+import com.liferay.application.list.BasePanelCategory;
+import com.liferay.application.list.PanelCategory;
 import com.liferay.application.list.constants.PanelCategoryKeys;
-import com.liferay.portal.service.PortletLocalService;
-import com.liferay.portal.util.PortletKeys;
+import com.liferay.portal.kernel.language.LanguageUtil;
+
+import java.util.Locale;
 
 import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 
 /**
  * @author Eudaldo Alonso
@@ -29,24 +29,26 @@ import org.osgi.service.component.annotations.Reference;
 @Component(
 	immediate = true,
 	property = {
-		"panel.category.key=" + PanelCategoryKeys.CONTROL_PANEL_CONFIGURATION,
-		"service.ranking:Integer=300"
+		"panel.category.key=" + PanelCategoryKeys.CONTROL_PANEL,
+		"service.ranking:Integer=400"
 	},
-	service = PanelApp.class
+	service = PanelCategory.class
 )
-public class ServerAdministrationPanelApp
-	extends BaseControlPanelEntryPanelApp {
+public class ConfigurationPanelCategory extends BasePanelCategory {
 
 	@Override
-	public String getPortletId() {
-		return PortletKeys.ADMIN_SERVER;
+	public String getIconCssClass() {
+		return "icon-cog";
 	}
 
-	@Reference(unbind = "-")
-	protected void setPortletLocalService(
-		PortletLocalService portletLocalService) {
+	@Override
+	public String getKey() {
+		return PanelCategoryKeys.CONTROL_PANEL_CONFIGURATION;
+	}
 
-		this.portletLocalService = portletLocalService;
+	@Override
+	public String getLabel(Locale locale) {
+		return LanguageUtil.get(locale, "category.configuration");
 	}
 
 }
